+++ CHANGELOG.md
@@ -96,6 +96,7 @@
 - PROJECT : Add new feature 'Project template'
 - Contract : Add analytic information to contract lines.
 - MESSAGE TEMPLATE : help to suggest use of separator between email addresses.
+- PRODUCTION ORDER : user can define manuf order's planned end date while creating production order from product form
 
 ## Improvements
 - Contract: added button to manually close contract if termination date was set in the future